requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
PyMySQL==1.1.1
cryptography>=42.0.0
Werkzeug==3.0.3
Pillow>=10.0.0
python-dateutil>=2.8.0
openpyxl>=3.0.0
